+++ semantic-conventions/src/opentelemetry/semconv/model/semantic_attribute.py
@@ -27,6 +27,9 @@ from opentelemetry.semconv.model.utils import (
     validate_values,
 )
 
+TEMPLATE_PREFIX = "template["
+TEMPLATE_SUFFIX = "]"
+
 
 class RequirementLevel(Enum):
     REQUIRED = 1
@@ -75,6 +78,10 @@ class SemanticAttribute:
         return replace(self, inherited=True)
 
     @property
+    def instantiated_type(self):
+        return AttributeType.get_instantiated_type(self.attr_type)
+
+    @property
     def is_local(self):
         return not self.imported and not self.inherited
 
@@ -117,7 +124,9 @@ class SemanticAttribute:
                 raise ValidationError.from_yaml_pos(position, msg)
             if attr_id is not None:
                 validate_id(attr_id, position_data["id"])
-                attr_type, brief, examples = SemanticAttribute.parse_id(attribute)
+                attr_type, brief, examples = SemanticAttribute.parse_attribute(
+                    attribute
+                )
                 if prefix:
                     fqn = f"{prefix}.{attr_id}"
                 else:
@@ -231,7 +240,7 @@ class SemanticAttribute:
         return attributes
 
     @staticmethod
-    def parse_id(attribute):
+    def parse_attribute(attribute):
         check_no_missing_keys(attribute, ["type", "brief"])
         attr_val = attribute["type"]
         try:
@@ -242,14 +251,9 @@ class SemanticAttribute:
             position = attribute.lc.data["type"]
             raise ValidationError.from_yaml_pos(position, e.message) from e
         brief = attribute["brief"]
-        zlass = (
-            AttributeType.type_mapper(attr_type)
-            if isinstance(attr_type, str)
-            else "enum"
-        )
-
         examples = attribute.get("examples")
         is_simple_type = AttributeType.is_simple_type(attr_type)
+        is_template_type = AttributeType.is_template_type(attr_type)
         # if we are an array, examples must already be an array
         if (
             is_simple_type
@@ -263,18 +267,31 @@ class SemanticAttribute:
             # TODO: If validation fails later, this will crash when trying to access position data
             # since a list, contrary to a CommentedSeq, does not have position data
             examples = [examples]
-        if is_simple_type and attr_type not in (
-            "boolean",
-            "boolean[]",
-            "int",
-            "int[]",
-            "double",
-            "double[]",
+        if is_template_type or (
+            is_simple_type
+            and attr_type
+            not in (
+                "boolean",
+                "boolean[]",
+                "int",
+                "int[]",
+                "double",
+                "double[]",
+            )
         ):
             if not examples:
                 position = attribute.lc.data[list(attribute)[0]]
                 msg = f"Empty examples for {attr_type} are not allowed"
                 raise ValidationError.from_yaml_pos(position, msg)
+
+        if is_template_type:
+            return attr_type, str(brief), examples
+
+        zlass = (
+            AttributeType.type_mapper(attr_type)
+            if isinstance(attr_type, str)
+            else "enum"
+        )
 
         # TODO: Implement type check for enum examples or forbid them
         if examples is not None and is_simple_type:
@@ -359,6 +376,29 @@ class AttributeType:
         )
 
     @staticmethod
+    def is_template_type(attr_type: str):
+        if not isinstance(attr_type, str):
+            return False
+
+        return (
+            attr_type.startswith(TEMPLATE_PREFIX)
+            and attr_type.endswith(TEMPLATE_SUFFIX)
+            and AttributeType.is_simple_type(
+                attr_type[len(TEMPLATE_PREFIX) : len(attr_type) - len(TEMPLATE_SUFFIX)]
+            )
+        )
+
+    @staticmethod
+    def get_instantiated_type(attr_type: str):
+        if AttributeType.is_template_type(attr_type):
+            return attr_type[
+                len(TEMPLATE_PREFIX) : len(attr_type) - len(TEMPLATE_SUFFIX)
+            ]
+        if AttributeType.is_simple_type(attr_type):
+            return attr_type
+        return "enum"
+
+    @staticmethod
     def type_mapper(attr_type: str):
         type_mapper = {
             "int": int,
@@ -432,7 +472,9 @@ class EnumAttributeType:
         otherwise it returns the basic type as string.
         """
         if isinstance(attribute_type, str):
-            if AttributeType.is_simple_type(attribute_type):
+            if AttributeType.is_simple_type(
+                attribute_type
+            ) or AttributeType.is_template_type(attribute_type):
                 return attribute_type
             # Wrong type used - raise the exception and fill the missing data in the parent
             raise ValidationError(