+++ semantic-conventions/src/opentelemetry/semconv/templating/markdown/__init__.py
@@ -22,6 +22,7 @@ from pathlib import PurePath
 
 from opentelemetry.semconv.model.constraints import AnyOf, Include
 from opentelemetry.semconv.model.semantic_attribute import (
+    AttributeType,
     EnumAttributeType,
     EnumMember,
     RequirementLevel,
@@ -97,11 +98,11 @@ class MarkdownRenderer:
         """
         This method renders attributes as markdown table entry
         """
-        name = self.render_attribute_id(attribute.fqn)
+        name = self.render_fqn_for_attribute(attribute)
         attr_type = (
             "enum"
             if isinstance(attribute.attr_type, EnumAttributeType)
-            else attribute.attr_type
+            else AttributeType.get_instantiated_type(attribute.attr_type)
         )
         description = ""
         if attribute.deprecated and self.options.enable_deprecated:
@@ -184,7 +185,8 @@ class MarkdownRenderer:
     ):
         attr_to_print = []
         for attr in sorted(
-            semconv.attributes, key=lambda a: "" if a.ref is None else a.ref
+            semconv.attributes_and_templates,
+            key=lambda a: "" if a.ref is None else a.ref,
         ):
             if self.render_ctx.group_key is not None:
                 if attr.tag == self.render_ctx.group_key:
@@ -192,6 +194,7 @@ class MarkdownRenderer:
                 continue
             if self.render_ctx.is_full or attr.is_local:
                 attr_to_print.append(attr)
+
         if self.render_ctx.group_key is not None and not attr_to_print:
             raise ValueError(
                 f"No attributes retained for '{semconv.semconv_id}' filtering by '{self.render_ctx.group_key}'"
@@ -275,7 +278,7 @@ class MarkdownRenderer:
             )
 
             for attr in sampling_relevant_attrs:
-                output.write("* " + self.render_attribute_id(attr.fqn) + "\n")
+                output.write("* " + self.render_fqn_for_attribute(attr) + "\n")
 
     @staticmethod
     def to_markdown_unit_table(members, output: io.StringIO):
@@ -325,19 +328,37 @@ class MarkdownRenderer:
             if notes:
                 output.write("\n")
 
+    def render_fqn_for_attribute(self, attribute):
+        rel_path = self.get_attr_reference_relative_path(attribute.fqn)
+        name = attribute.fqn
+        if AttributeType.is_template_type(attribute.attr_type):
+            name = f"{attribute.fqn}.<key>"
+
+        if rel_path is not None:
+            return f"[`{name}`]({rel_path})"
+        return f"`{name}`"
+
     def render_attribute_id(self, attribute_id):
         """
         Method to render in markdown an attribute id. If the id points to an attribute in another rendered table, a
         markdown link is introduced.
         """
+        rel_path = self.get_attr_reference_relative_path(attribute_id)
+        if rel_path is not None:
+            return f"[`{attribute_id}`]({rel_path})"
+        return f"`{attribute_id}`"
+
+    def get_attr_reference_relative_path(self, attribute_id):
         md_file = self.filename_for_attr_fqn.get(attribute_id)
         if md_file:
             path = PurePath(self.render_ctx.current_md)
             if path.as_posix() != PurePath(md_file).as_posix():
-                diff = PurePath(os.path.relpath(md_file, start=path.parent)).as_posix()
-                if diff != ".":
-                    return f"[`{attribute_id}`]({diff})"
-        return f"`{attribute_id}`"
+                rel_path = PurePath(
+                    os.path.relpath(md_file, start=path.parent)
+                ).as_posix()
+                if rel_path != ".":
+                    return rel_path
+        return None
 
     def to_markdown_constraint(
         self,
@@ -390,7 +411,9 @@ class MarkdownRenderer:
                         )
                     a: SemanticAttribute
                     valid_attr = (
-                        a for a in semconv.attributes if a.is_local and not a.ref
+                        a
+                        for a in semconv.attributes_and_templates
+                        if a.is_local and not a.ref
                     )
                     for attr in valid_attr:
                         m[attr.fqn] = md